requirements.txt
## GeoAI Flood Risk System – Requirements

# Core Geospatial
rasterio>=1.3
numpy>=1.26
scipy>=1.11
pyproj>=3.5

# Machine Learning
torch>=2.0
scikit-learn>=1.3

# Web Framework
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0

# Visualization & Frontend Tools
Pillow>=10.0
matplotlib>=3.7
pandas>=2.0

# Dashboard (Streamlit kept for backward compatibility with old app.py)
streamlit>=1.30
streamlit-folium>=0.16
folium>=0.15

# Network requests (live weather endpoint)
requests>=2.31
